requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
certifi==2017.7.27.1      # via requests
chardet==3.0.4            # via requests
click==6.7                # via flask
flask==0.12.2             # via locustio
gevent==1.2.2             # via locustio
greenlet==0.4.12          # via gevent
idna==2.6                 # via requests
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
locustio==0.8.1
markupsafe==1.0           # via jinja2
msgpack-python==0.4.8     # via locustio
pyzmq==16.0.2             # via locustio
requests==2.18.4          # via locustio
six==1.11.0               # via locustio
urllib3==1.22             # via requests
werkzeug==0.12.2          # via flask
